Improved version of previous fixes
Allow messages with 0 domain data size to be handled and passed
through various routines. These can be (and are) sent when the
recipient will need metainformation (aka tree data) abour a remote
block, but not ant payload data from that block.  Using this, he
communication savings of GRID_SUBPAT_GC_OPT are now less reduced than
in the previous variant.