comment out (ifdef MAYBE_UNNECESSARY) nlayers-increasing code
The code in question was kicking in for cases of large
nguard-to-n[xyz]b ratios. I am disabling it by default (unless you
(1) it is unclear how this helped at all
    (maybe it is a leftover from old PARAMESH versions that had dedicated
    f_to_c routines for guard cell filling at f/c boundaries);
(2) nxb/nguard < 2 etc. should be avoided anyway for various reasons
    (a) may fill invalid data in coarse blocks where f-to-c
        restriction is necessary;
    (b) things can be expected to COMPLETELY break down anyway if
        (nguard > any {nxb,nyb,nzb}) is true;
    (c) It is highly inefficient / wasteful of memory to have high
    nguard-to-{{nxb,nyb,nzb} ratios;
(3) for valid operations, this code needs to be enabled (or disabled)
    consistently between the two source files, which is rather ugly and
    error-prone.